CouponService implementation with controller refactor

+++ src/main/java/com/rataz/investmenttracking/coupon/Coupon.java
@@ -1,7 +1,6 @@
 package com.rataz.investmenttracking.coupon;
 
 
-
 import com.rataz.investmenttracking.event.Event;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -15,18 +14,18 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 @Entity
-@Table(name="coupon")
+@Table(name = "coupon")
 public class Coupon {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
-    @Column(name="is_won")
+    @Column(name = "is_won")
     private boolean isWon;
     private BigDecimal odds;
     private BigDecimal stake;
 
-    @OneToMany(mappedBy = "coupon",cascade = CascadeType.ALL,fetch = FetchType.LAZY)
+    @OneToMany(mappedBy = "coupon", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
     private List<Event> events;
 }
