+++ src/main/java/com/rataz/investmenttracking/coupon/CouponRepository.java
@@ -2,11 +2,8 @@ package com.rataz.investmenttracking.coupon;
 
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 
-import java.util.List;
-
-public interface CouponRepository  extends JpaRepository<Coupon,Integer> {
+public interface CouponRepository extends JpaRepository<Coupon, Long> {
 
 //    @Query( "SELECT c FROM Coupon c")
 //    List<Coupon> getAllCoupons();
