+++ src/main/java/com/rataz/investmenttracking/event/details/result/EventResultDetailRepository.java
@@ -2,5 +2,5 @@ package com.rataz.investmenttracking.event.details.result;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventResultDetailRepository extends JpaRepository<EventResultDetail,Integer> {
+public interface EventResultDetailRepository extends JpaRepository<EventResultDetail, Long> {
 }
