+++ src/main/java/com/rataz/investmenttracking/event/details/statistic/EventStatisticDetailRepository.java
@@ -2,5 +2,5 @@ package com.rataz.investmenttracking.event.details.statistic;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventStatisticDetailRepository extends JpaRepository<EventStatisticDetail,Integer> {
+public interface EventStatisticDetailRepository extends JpaRepository<EventStatisticDetail, Long> {
 }
